global: Change tournament enemy_count to uint32_t
This is because pyomftools has tournament enemy_count as uint32_t.
Hopefully fixes a static analysis complaint.

+++ include/formats/tournament.h
@@ -54,7 +54,7 @@ typedef struct {
  * Tournament enemies, locales, quotes, name, etc.
  */
 typedef struct {
-    int16_t enemy_count;            ///< Number of enemies in tournament
+    uint32_t enemy_count;           ///< Number of enemies in tournament
     char bk_name[14];               ///< Tournament BK filename
     float winnings_multiplier;      ///< Match winnings multiplier
     int32_t unknown_a;              ///< Unknown /todo find out
